Buffet server: add support for both devices and commands.
This CL adds both the device and commands methods to buffet's
fake_device_server.

For devices:
- Devices are created when either registration tickets are finalized
OR a device INSERT method (via a POST) is called.

For commands:
- Commands are associated to devices by deviceId which is required
whenever creating a new command. I store the mapping between devices
and their commands in a dictionary. The only reason I need to keep a
mapping is I need to map a device to its commands for the listing
method i.e. GET on commands without an id.

I have also made a small common class used by devices/commands that
implements PATCH/PUT as they are indentical for both. I almost did the
same for registration but I couldn't find a good way to deal with the
"claim" part of it.

I have left a couple pieces left TODO as they will require a bit more
thinking and we won't likely need them immediately. These are:
1) The upload API i.e. commands that require uploading a blob of data
to give to a device e.g. storage.writeFile.
2) A couple specific fields like owner in devices which don't seem
to matter yet.

BUG=chromium:358277
TEST=See README + unittests

Change-Id: I0eb11a2f1bdb31b5d3dc4f17a72499f5160eafc2
Reviewed-on: https://chromium-review.googlesource.com/193529
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>