Fix stable states list and provision reboot test
Changes include things discussed in the email thread and
listed in the attached bug. In brief, adding DISABLED states to stable
states and removing CONNECTING states from stable states. As well as
adding a flag for modems that may not auto register.

Also added the needed hooks to run this reboot test before
network_3GModemControl.e362 test. I have looked at the test results for
this test and it looks like a good candidate to try the reboot via
provisioning task. If this works well, we will add it to rest of the
cellular tests.

BUG=chromium:360140
TEST=Tested locally and on few machines in the lab.

Change-Id: Ibb8213dbcd5022e1ccde2289b5bfa24e181a065d
Reviewed-on: https://chromium-review.googlesource.com/193217
Reviewed-by: harpreet Grewal <harpreet@chromium.org>
Tested-by: harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Kris Rambish <krisr@chromium.org>
Commit-Queue: harpreet Grewal <harpreet@chromium.org>