Chameleon: Add the EDID files to the test
The test data, which currently contains EDID files, for the Chameleon test
display_Resolution, was compressed as a zipped tar file and put in GS server.
An ebuild is to install the test data to the test directory while running
build_package.

The lab Autotest server does not emerge this ebuild and the test data is not
installed on lab run. So we are going to remove the ebuild and add the test
data to the test.

BUG=chromium:359483
TEST=test_that --args "chameleon_host=$CHAMELEON" $DUT display_Resolution.*
CQ-DEPEND=I6932228a986b8e92db1b0c0ac00bd81699212fca

Change-Id: Iee79c9c05f5983d58b6f958fbdc951a934362a59
Reviewed-on: https://chromium-review.googlesource.com/193562
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>