[autotest] Treat special task labels like "provisionable" labels.
As it stands, to use the special task labels, you need to add the label
to every host, since we still treat it as a normal/standard DEPENDENCY.

If any special task treats a label differently, then we shouldn't
require that label on the DUT, as we're using it as a marker that we
want some special handling of this test rather than an actual
dependency.

BUG=chromium:334418
DEPLOY=scheduler, apache
TEST=added a potato->dummy_PassServer mapping, saw dummy_PassServer run
in cleanup on job with DEPENDENCIES=potato where host didn't have
potato.

Change-Id: I34936a5a84a42897c8dfc0e2acd50195e0e74ec5
Reviewed-on: https://chromium-review.googlesource.com/194212
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>
Tested-by: Alex Miller <milleral@chromium.org>