Adjust finger position to prevent collisions
Since the robot used relative coordinates that scale based on the size
of the touchpad being tested, for smaller touchpads the fingers can
sometimes bump into eachother.  This causes a weird bend in the logs and
frankly is just not what we're going for.

This CL adds in some logic to the touch robot test that trying to take
into consideration the size of the touchpad before proceeding and
adjusts the stationary finger's position if it's too close/far away.

BUG=chromium:354593
TEST=manual testing

Change-Id: If2d3ee3da7a37e902c0a06f69ea6cdb6395526cf
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/198420
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>