[autotest] Move power to pool:suites.
I can't seem to remember any reason why pool:power existed in the first
place?  We dont seem to be keeping up with adding new devices to this
pool anyway, so we might as well merge it with pool:suites to save us
the troubles.

BUG=None
TEST=sanity

Change-Id: I15fed63baae5f6195f3fd8ba0442fcd2981fc246
Reviewed-on: https://chromium-review.googlesource.com/194382
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>