Chameleon: Add the 'chameleon' label for hosts with Chameleon boards
This change adds a label 'chameleon' to represent if the host connects
with a Chameleon board. The Chameleon tests have the dependency on the
label chameleon. The label is detemined by the conditions:
 - the host is in Cros Lab;
 - specified by the argument chameleon_host.

BUG=none
TEST=Manual
To test that the label is assigned correctly, modified the contol file
of display_Resolution.mirrored:

def run(machine):
    host = hosts.create_host(machine, chameleon_args=chameleon_args)
    labels = host.get_labels()
    print '\n\n\nXXX\n'
    print labels

Ran the test display_Resolution.mirrored on a lab machine with a Chameleon.
Ran the same test specifying the argument chameleon_host.
Ran the same test without specifying the argument chameleon_host.

Change-Id: If64af00eee1e5d74f8eedcdf5dbe4a03dc546117
Reviewed-on: https://chromium-review.googlesource.com/194506
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>