autotest (dhcp): add test for NAK handling
This test sends NAK messages, without a server-ID, and
verifies that the DHCP client restarts DHCP negotiation.

We test two scenarios: a) the DHCP client is starting up
(but with a cached lease), and b) an already running DHCP
client is asked to renew its current lease.

BUG=chrome-os-partner:27930
TEST=manual

Manual test: ran this test before and after patching dhcpcd.
Test failed before the change, and passed afterwards. (It
took a couple iterations of this to get things right.)

Change-Id: I37329927e600568bc1268a1bfe54cfb17d9c5d5d
Reviewed-on: https://chromium-review.googlesource.com/194944
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>