Chameleon: Check heath and repair Chameleon if needed
Chameleon supports repair routines to recover itself. On the cleanup phase
of a test, check its health and repair it if needed, such that the next
test cases are able to be ran.

CQ-DEPEND=CL:195087
BUG=chromium:360993
TEST=Manaul
Together with the CL of CQ-DEPEND. Ran Chameleon test suite on lab machines.
  $ test_that --fast $LAB_HOST suite:chameleon
Saw the log that the repair rountine was executed.

Change-Id: I80de19ffbd9a364043afa3d7cb6b2558c4f20216
Reviewed-on: https://chromium-review.googlesource.com/195091
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>