Chameleon: Increase the timeout margins of suspend and resume
The purpose of the Chameleon test is to check the display still correct
after suspend and resume. The suspend time and the resume time fitting the
requirements is not what we care. So increasing the timeout margins to
reduce the false alarms, which are not belonging to the display test.

BUG=chromium:359555
TEST=Ran the display_HotPlugAtSuspend test >50 times without failure.

Change-Id: I2c48c6e11b4475c001eb0129ff351c357f61ae02
Reviewed-on: https://chromium-review.googlesource.com/195338
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>