autotest (platform): add test for syslog duplicate suppression
We've had a couple of bugs recently, where processes filled up the
disk with log messages. In both cases, a single message was repeated
over and over, thousands of times per second.

Fortunately, there's a rsyslogd configuration option that deals
with exactly this case. I'm turning that on in CL:195513, and this
test verifies that the suppression works.

BUG=chromium:363224
TEST=manual

Manual test: Ran this test before and after CL:195513. Verified that
the test failed before CL:195513, and passed after.

Change-Id: Idc5b4ed36c30c19b21f87fa843320c869d3360c9
Reviewed-on: https://chromium-review.googlesource.com/195517
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>