[autotest] Move autoupdate_Rollback to AU suite.
I'm receiving complaints about autoupdate_Rollback failing when people
run BVT from their desk, as there's a reliance on a devserver to run the
test.  The other tests that rely on infrastructure are in the au suite,
so we can keep autoupdate_Rollback as part of the waterfall by just
shuffling it over.

BUG=chromium:364952
TEST=None

Change-Id: I00b475286b567737f31d56982b4ae916bf8091ad
Reviewed-on: https://chromium-review.googlesource.com/195687
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>