Touch FW test: handle missing tracking id's
Sometimes the tracking ID's weren't getting picked up during the period
of time the logging was running and that crashes the parser.  This patch
adds a check it to prevent this.  If events are coming in for a finger
that we don't have a tid for yet, it simply assigns it the dummy value
999999.  This should allow the tests to work and parse even if the log
misses the first bit where the tracking ID is assigned by FW.

BUG=chromium:367940
TEST=I used this to parse a problematic log that was crashing it before.
It seems to work now.  Also ran the unit tests without any failures now

Change-Id: I3f4a95a3e6b19fcd1992f38db23dfe69289e953d
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/197380
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>