Create the 'experimental' suite for experimental tests.
Use of EXPERIMENTAL = "True" is being deprecated.  Instead,
"experimental" tests are to be place in any non-BVT suite.  This
provides a new suite named "experimental" that runs on each new
build as a convenience for tests that have no other suite to live
in.

BUG=None
TEST=None

Change-Id: Idd1734098c47ed391fb94a3764aae85f8b335b1b
Reviewed-on: https://chromium-review.googlesource.com/197379
Reviewed-by: Prashanth B <beeps@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>