[autotest] Do not log servo reboot during servo update
If servo reboot is logged, i.e., we wait for reboot to complete, a failed reboot
will leave a failure record in status.log, and that leads to test job failure
even if the test is already finished and does not require servo.

The fix is to reboot servo without waiting in self.reboot call, thus avoid
logging the reboot record.

BUG=chromium:367997
TEST=None
The local setup with test network makes it impossible to reboot a local servo
without hacking the code.

Change-Id: Iaab2b5c09a5793ba7c6a757cc16c1291cc433baa
Reviewed-on: https://chromium-review.googlesource.com/197526
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>
Tested-by: Dan Shi <dshi@chromium.org>