[autotest] Schedule provision only if there is a provisionable label.
Adding actionable labels (labels that verify, cleanup, etc. can do
special things with) broke an assumption that provisioning code had that
the only time we'd have labels on a job that weren't on the host were if
they were provisionable.  This fixes that assumption to check if they
are or are not provisionable.

BUG=chromium:369079
TEST=Ran a job through the scheduler that had an actionable but
non-provisionable label.

Change-Id: I7255ec3368cbe3ee846ec783787e91efd8ba9d59
Reviewed-on: https://chromium-review.googlesource.com/198048
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>