Spreadsheet: Automation support
This CL adds command line options to set values that would otherwise be
asked for interactively.
It also adds an option to print the results of the calibration process
by reading it from the spreadsheet.
With these two changes it is possible to run the pressure calibrations
script programmatically and retreive the final calibration info. This
will be used by the MTBringup tool to automate device bringup tasks.

BUG=chromium:369610
TEST=run spreadsheet.py --print-info --name "Glimmer"
should print calib info for Glimmer

Change-Id: Ife881aab71009fdfd4bbc022cb5d2e3a8a32c634
Reviewed-on: https://chromium-review.googlesource.com/198019
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@chromium.org>
Commit-Queue: Dennis Kempin <denniskempin@chromium.org>
Tested-by: Dennis Kempin <denniskempin@chromium.org>