[autotest] Don't raise TestNA in provision tests.
Autoserv has this annoying distinction between the exit code of autoserv
and the result of status.log being parsed.  Raising a TestNA gets the
correct line into the status.log to show that the test was actually a
no-op, however, the fact that it's an exception makes autoserv think
that the test failed.  This means that cros-version provisioning will
always fail for a DUT that's already in the right state, which is not
what provisioning was meant to do.

Instead, we emit an INFO line to still try to communicate the
TEST_NA-ness of running the test, and then return normally so that we
end with a GOOD and END GOOD, and autoserv thinks the test passed.

BUG=chromium:369079
TEST=Ran autoserv --provision with cros-version label set to what was on
the host, autoserv returned 0.

Change-Id: Ifdd9a4286cb5756d5ddf7143ce6260a5b6d48210
Reviewed-on: https://chromium-review.googlesource.com/198049
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>