hardware_MemoryThroughput change to use lmbench
This changes the hardware_MemoryThroughput test to use lmbench and
removes the old code.  As far as I know nobody was paying attention to
the old version of the test and the results from it made little sense
to me, so I'm changing it to do something that makes more sense.

BUG=chromium:354161
TEST=`test_that <ip> hardware_MemoryThroughput`

Change-Id: I03c6f1368ecf30f3b88c9cea8ef203bc08f1f72c
Reviewed-on: https://chromium-review.googlesource.com/198765
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Queue: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>