[autotest] models.Job set default synch_count to 0.
Currently when creating a suite job we pass in a synch_count of
None, pass synch_count validation, then create a Job entry in
the database. However the default value for synch_count is 1 and
that is what gets recorded in the database.

This breaks the ability to create a recurring job for a suite job
as the original job's synch_count does not match what was used to
create it. In order to address this, make the default value for
synch_count be 0.

BUG=chromium:369726
TEST=./database/migrate sync and ensured afe_jobs now shows not null
for synch_count. Created a regular job from the afe, kicked off
a suite via run_suite, created a recurring suite job and a recurring
regular job. frontend_unittest.py
DEPLOY=scheduler,apache,migrate

Change-Id: I31728f9f876c917a3acaa390c7b9ffc33042bbc2
Reviewed-on: https://chromium-review.googlesource.com/198102
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>