firmware_TouchMTB: upgrade to the new Google spreadsheet
We have manually upgraded the old Google spreadsheet to the new one.
Here we need to use the new key to access the new spreadsheet.
Because CONTINUE function is deprecated in the new style, we use
transpose function to achieve the same purpose.

It looks that the table operations in a worksheet is not supported
yet. We skip the table operations as a temporary work around.

BUG=chromium:363384
TEST=Follow the instructions on the Touchpad Pressure Calibration Guide:
https://docs.google.com/a/google.com/document/d/1XOL7M0Pw-tlN3Z7UVWgfbkkQz6CiXMbnhO2CL3LankU/edit

It should be able to insert a new worksheet successfully.

Change-Id: Ia53ddb2b1ddb44804055a0278b5c8d184dd3e3cc
Reviewed-on: https://chromium-review.googlesource.com/198148
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>