Add the reset script back into touchbotII wrapper
Somehow the reset script got removed from the touchbotII wrapper so the
robot was calling it a little close when returning to the touchpad of
the device it's testing.  This CL adds it back in, so now after getting
new fingertips the robot will lift up and away before trying to run the
next gesture to reduce the likelyhood of collisions.

BUG=none
TEST=I just used this version for a test and it worked great

Change-Id: I256862bcba978994ed80b9c8a7733b6820659bc3
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/198195
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>