Explicitly disconnect after beacon failure
In the network_WiFi_MissingBeacons test, we ensure a Service
enters the "idle" state, and immediately start another connection
attempt.  It's possible that shill may be re-trying the recently
failed connection at the same time as the test is passively trying
to detect the new AP.  Fix this by explicitly asking shill to
disconnect from the old AP at the end of _assert_disconnect.

BUG=chromium:370156
TEST=Run test on system that originally encounered this problem

Change-Id: I741a4686a6fd16dd8442c54aaf28ec7ce9ac7b69
Reviewed-on: https://chromium-review.googlesource.com/198228
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>