Catching exception when checking for existance of firmware programmer utilities.
Ignoring the exception that occurs when checking for the existance of some
required utilties for the fw programmer.

Currently the firmware programmer is not working or used, and the check is
blocking us from running tests in environments that do not package the required
dependent bins.

We should reinstate this exception once the programmer is working to indicate
the missing utilities earlier in the test cycle, rather than hitting the
exception when they are called.

BUG=chromium:370288,chromium:371011
TEST=Manually at desk with/without require bins, and saw the expected msg.

Change-Id: Idbfbe1a43f65e0cda22a7f7c8814f431dfa6f124
Reviewed-on: https://chromium-review.googlesource.com/198307
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Yusuf Mohsinally <mohsinally@chromium.org>
Commit-Queue: Yusuf Mohsinally <mohsinally@chromium.org>
Tested-by: Yusuf Mohsinally <mohsinally@chromium.org>