power_LoadTest: Work around extension loading twice
This works around Chrome loading the extension for OOBE and after login by
making the extension not automatically start on load. The icon can still be
clicked to start the extension though.

Instead, this relies on power_LoadTest.py to start the test by executing a
javascript function for the extension.

BUG=chrome-os-partner:27981
TEST=run power_LoadTest
	check that only one set of windows opens

Change-Id: Ib10ce757b3fdabca4604f148374f80993fc34e85
Reviewed-on: https://chromium-review.googlesource.com/198278
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>