hardware_Trackpd: fix subset issue for localhost
When running the test on the gesture files captured locally
in /var/tmp, it needs to be assigned a subset so that the
test flow would not be interrupted as an exception. This simple
patch fixed the problem by assigning a "long" subset so that it
will run through all gestures in the folder.

BUG=chromium:370732
TEST=This is one of the patches in series.
Will be tested in a subsequent patch.

Change-Id: Ic43e8fd114031b53989f1abfd8a3bb4655af0d2e
Reviewed-on: https://chromium-review.googlesource.com/198608
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>