Add file utilities to support screenshot capturing.
Adds a file_utils.py file which provides methods to create and remove file/dirs.
These are wrappers around python utilies, however, the create method for example,
will not fail if the file/directory already exists.

BUG=chromium:363815
TEST=Ran local tests that use file_utils.py. Need to run existing tests that use site_utils.py

Adds a method in site_utils.py to take a cropped screenshots. We are only cropping
by height, adjusting width to retain aspect ratio.
We don't need a crop by width method at this time and so we do not provide it with this change.

Change-Id: I9f0943dc518c7ad96044d0a314c4422bba0a087f
Reviewed-on: https://chromium-review.googlesource.com/198395
Reviewed-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Commit-Queue: Rohit Makasana <rohitbm@chromium.org>
Tested-by: Rohit Makasana <rohitbm@chromium.org>