Add method to re-claim APs that are locked.
Sometimes the DUT will get in a bad state, see crbug.com/331915, and
the only fix is to reboot.  However once APs are added to the batch
they are removed from the the complete list of APs to process.

This new method, unlocks the remaining APs in the batch and adds them
back to the complete list.

BUG=None
TEST=Run connect/disconnect.

Change-Id: If2a1c8e6b4e91c1378fabc513611f4a1d9766355
Reviewed-on: https://chromium-review.googlesource.com/198600
Commit-Queue: Kris Rambish <krisr@chromium.org>
Tested-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Deepak Gopal <deepakg@chromium.org>