hardware_Trackpd: use "xev -geometry" to listen to its own window
When xev listens to mtplot to capture X events, the Button tap events
might be intercepted by mtplot and are not propagated to xev. This
will result in the test failure about tapping gestures.

This patch uses xev with the "-geometry" flag to create its own
window to capture the X events. In this way, all X events could be
captured by xev correctly.

However, we would also like to keep the flexibility of using mtplot
here. The advantage of using mtplot is that mtplot will display the
gestures on the window while the xev's own window will only be a
blank one. When there is a request of using mtplot later, it would
be easier to fix it to propagate all X events.

BUG=chromium:370732
TEST=Use one of the following machines that we have already had
gesture files recorded for the regression test:
Machines supported: mario, zgb, alex, and lumpy.

Step 1: Install hardware_Trackpad from chroot.
(cr) $ run_remote_tests.sh --remote=172.30.210.80 hardware_Trackpad/control$

Step 2: Conduct a regression test on the machine.
$ cd /usr/local/autotest/tests/hardware_Trackpad
$ ../../bin/autotest control

It should generate a test report to describe how many tests passed.

Change-Id: I561df55bd11574b3f52dd1668ed3a4d7d07414ee
Reviewed-on: https://chromium-review.googlesource.com/198642
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>