kernel_Delay: Fix cpu frequency setting issues on link
link has independent setting of CPU frequencies which was causing
problems because the CPU frequency wasn't getting set on all CPUs
which was manifesting as failures as the frequency setting didn't stick.

Move to a userspace governor using scaling_setspeed on each CPU.
Setting scaling parameters on each CPU causes issues if CPUs got
un-hotplugged (ie nyan), so move to userspace cpuquiet during the test
if the option is available to prevent that.

BUG=chromium:352391
TEST=test_that kernel_Delay on nyan, link, squawks, zako, spring, peach_pit

Change-Id: Ib6a4842ef68308c7434af58e42efc70d1e4cb746
Reviewed-on: https://chromium-review.googlesource.com/199802
Commit-Queue: David Riley <davidriley@chromium.org>
Tested-by: David Riley <davidriley@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>