power_CheckAC: New test to be used for RPM verification
Add a test to check that the line power is in the expected state. There doesn't
seem to be a clean way to check the line status on a remote machine and
power_status already has one workaround for butterfly's line status bug.

BUG=chromium:368466
TEST=run power_CheckAC when plugged into AC
CQ-DEPEND=CL:199640

Change-Id: I6b2f93a69f4470121c82ead819c0c29665c5ef11
Reviewed-on: https://chromium-review.googlesource.com/198776
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>