Add auth support to client_lib and test to be able to use against the live server.
In this CL I've added various oauth helpers to convert codes to auth tokens
for use when interacting with our actual servers. In order to facilitate quick
testing I've modified this test script to either test against the fake
server OR the live server based on whether you pass in an auth code
to the command line (see usage).

There's still a couple issues (left as TODO's here) but this seems like
a good point for review.

BUG=chromium:361204
TEST=Tested up to command insertion with live server + fake server + unittests

Change-Id: Ic7166ddaa50d171dbf9a8cbb4b50b736f79a32b6
Reviewed-on: https://chromium-review.googlesource.com/198784
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>