Bump cros_stable_version to R36-5832.0.0.
nyan_big systems in the lab are failing to repair because the
default stable image is too old to work:  If you install that
current image on nyan_big the kernel crashes and rolls back.

Bumping the stable image version to R36-5832.0.0; that build
passed BVT on all boards except nyan_big (where it didn't run),
and nyan_big had passed on earlier builds.

BUG=None
TEST=None

Change-Id: I88cfe524654bc4d5b7c0fe852103e44d8e2a46a7
Reviewed-on: https://chromium-review.googlesource.com/198788
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>