Fix hangs in TouchMTB FW test on no gestures
Previously, the timeout callback that wraps up a gesture in the Touch FW
test is only activated when the DUT sees the first event come in from
the touchpad.  This presents a problem in two situations:

  1.) The robot decides it is incapable of a gesture.  This is possible
      for some of the two-fingered gestures depending on the shape of
      the touchpad.  Obviously, it's be better if it never happened, but
      it will happen sometimes and we don't want it to crash when it
      does.

  2.) There is a test (such as detecting taps) where the failure mode is
      that no events are ever reported by the touchpad.  In this
      situation, even though the robot did what it was supposed to, the
      DUT never sees anything and so it just hangs.

with this change both kinds of failures are simply ignored.  A blank
entry is added to the html report and it's skipped when calculating
the summary scores.

For case number 1, this is the desired behavior.  (You shouldn't hold it
against the pad that the robot can't do the gesture)  However for case
number 2, this is in fact the fault of the touchpad and should be
reported as a failure.  This CL doesn't do it, but at least it keeps it
from hanging, so this is the first step.  I will work on a fix for #2
next.

BUG=chromium:371508
TEST=manual testing on a device that was exhibiting both of these
failure modes.  The failed tests responded as described above and
everything else worked as usual

Change-Id: I174b225c4375b8aa1d74292cd65c4653a27560ce
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/199148
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>