[autotest] Create one authoritative incremental way to clean a database
There should only be one way to clean the database, and the by date
script is more recent, so I'm considering it the right way.  It belongs
in contrib, until it actually gets fixed and cleaned up more, so it's
moving there, and db_clean is just going away.

BUG=chromium:372939
TEST=ran db_cleanup_by_date (now db_clean)

Change-Id: Id36e7336b61ca17637560fb4b65ca127016a7ae0
Reviewed-on: https://chromium-review.googlesource.com/199562
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>
Tested-by: Alex Miller <milleral@chromium.org>