graphics_Piglit: Don't fail test if no tests failed (or passed)
If the "fail" findall() returns an empty list, it means that 0 tests
failed.  This really shouldn't be considered a TestFail!

In fact, the current pass/fail criteria for the graphics_Piglit is whether
the test suite ran successfully.  It is not based on how many tests
passed or failed.  Therefore, even if 0 tests passed, it still shouldn't
be a TestFail.

v2: add empty summary check.
v3: raise TestError, not return

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:331326
TEST=manually modify test list to run just one test that is known to pass
TEST=./run_remote_tests.py graphics_Piglit
  => Test should pass
TEST=manually modify test list to run just one test that is known to fail
TEST=./run_remote_tests.py graphics_Piglit
  => Test should pass

Change-Id: I4ce4fa13ec37edbf293534a5a2fd8f32e4355a43
Reviewed-on: https://chromium-review.googlesource.com/184071
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>