Adds support to download golden images.
We need to know if the test images we captured are as expected. To do that
we need to have golden images. The current comparerer requires both the
golden image and the test image to reside locally on the device.

Thus, we need to download the golden images in the beginning of the test.

** Added error checking for image downloading.

BUG=chromium:363815
TEST=Ran local tests that use the downloader to download images. Observed
that downloaded images are correct.

Change-Id: I5dee61919d97bc6252802983bbd28b6fd761b372
Reviewed-on: https://chromium-review.googlesource.com/199426
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Reviewed-by: Kris Rambish <krisr@chromium.org>
Tested-by: Rohit Makasana <rohitbm@chromium.org>
Commit-Queue: Rohit Makasana <rohitbm@chromium.org>