Add a quickstep flag to the touch FW test
It's common to want to check the latency of a system without running the
full (long) FW test on it.  This CL adds a nice, simple way to do that
without having to know how the code works and comment out parts you
don't want.  Now there is an additional '-q' or '--quickstep' flag you
can pass to the test suite when you run it.  If you do it overrides the
list of gestures and only runs the latency test.

BUG=chromium:372895
TEST=I used it both with and without the flag and it seems to do the
job.

Change-Id: Ia05b771c00341447e253d56a13a7c08e967d8401
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/199495
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>