[autotest] Ensure test_suites is staged when using dynamic_suites
Currently we stage test_suites in the Apache RPC when creating a suite job.
This is because we don't have the control file and need to retrieve it to
create the job.

However if a suite job is cloned and the devserver has deleted this artifact,
control_file_getter will fail as the artifacts are now missing. Therefore
to address this stage both autotest and test_suites as part of
preform_reimage_and_run.

BUG=chromium:372593
TEST=Cloned a suite job and deleted the build artifacts beforehand. Suite ran
as normal.

Change-Id: Ib9bcc40ce8f71ba2873844fa428e25c89e71336a
Reviewed-on: https://chromium-review.googlesource.com/199680
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>