Touch FW Test: recover from dropped fingertips
Currently the robot will raise exceptions when it looses track of the
fingertips currently equipped.  This is a safe, but very frustrating
policy since it halts the whole test, and it must be restarted.

This CL adds some code that instead of raising exceptions in this case,
simple pauses the test until the operator manually configures the robot
and tells it that it's safe to continue.  This is slightly less safe,
since the operator could make a mistake, but if we're careful it should
be fine, and it should make things less frustrating to run.
Additionally, there are some scary warnings that pop up when this
happens so hopefully anyone who isn't 100% sure of what they're doing at
this point will contact someone who does.

BUG=chromium:373380
TEST=I forced the robot to get confused and it was handled gracefully

Change-Id: Ia2dcf05c47a521b5694a46681c7425e1e341764f
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/199798