Adds support to interact with Chrome HTML5 player.
This class provides easy methods to load a particular page into a browser tab,
wait for the page to be ready and also wait for interesting javascript events
such as 'canplay' and 'seeked'.

This is class is the means to control the HTML5 video player currently playing inside
a chrome tab.

Note that, the interface is designed to be generic for any player such as: Youtube player.

When that new player needs to be tested, we will build a class that implements the same
interface as this class so that we can take advantage of python's duck typing features
to inject it to the screenshot_collector class.

  ** Added video.html. media_player.py executes javascript functions defined in this file.

BUG=chromium:363815
TEST=Ran local tests that used this class to control browser.

Change-Id: Iaa3f9d9743b27a9d5735217e9b70d31a70a19d74
Reviewed-on: https://chromium-review.googlesource.com/199926
Reviewed-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Commit-Queue: Rohit Makasana <rohitbm@chromium.org>
Tested-by: Rohit Makasana <rohitbm@chromium.org>